reafactor: move commands to separate files
