+++ package.json
@@ -4,9 +4,10 @@
     "description": "Template for the WebSocket remote control task",
     "main": "index.js",
     "scripts": {
-        "start": "node ./index.js",
+        "start": "tsc && node dist/index.js",
         "start:dev": "nodemon ./index.ts",
-        "test": "echo \"Error: no test specified\" && exit 1"
+        "lint": "eslint \"src/**/*.{ts,js}\"",
+        "lint:fix": "npm run lint -- --fix"
     },
     "author": "Andrei Auchynnikau",
     "keywords": [],
