+++ src/ws_server/index.ts
@@ -1,23 +1,34 @@
-import { WebSocketServer } from 'ws';
+import { createWebSocketStream, WebSocketServer } from 'ws';
 import { COMMANDS } from './const';
 import { handler } from './handler';
 
-export const init = () => {
-    const wsServer = new WebSocketServer({ port: 8080 });
+const WS_PORT = 8080;
 
+export const wsServer = new WebSocketServer({ port: WS_PORT });
+
+export const initWsServer = () => {
     wsServer.on('connection', (ws) => {
         console.log('wsServer on connection');
 
-        ws.on('message', async (data) => {
+        const wsStream = createWebSocketStream(ws, {
+            decodeStrings: false,
+            encoding: 'utf8',
+        });
+
+        wsStream.on('data', async (data) => {
             const [command, ...params] = data.toString().split(' ');
             console.log(`Client command: command - ${command}, params - ${params}`);
 
             try {
                 const result = await handler(command as COMMANDS, params);
-                ws.send(result);
+                wsStream.write(result);
             } catch (error) {
-                console.log(`error ${error}`);
+                console.error(`error ${error}`);
             }
+        });
+
+        wsStream.on('error', (error) => {
+            console.error(error);
         });
     });
 
@@ -25,7 +36,7 @@ export const init = () => {
         console.log('wsServer on close');
     });
 
-    wsServer.on('error', () => {
-        console.log('wsServer on error');
+    wsServer.on('error', (error) => {
+        console.error(`wsServer on error ${error}`);
     });
 };
